Switched base.py; organized tests
