+++ tests/test_spatial_utils.py
@@ -8,7 +8,6 @@ import xarray as xr
 from numpy.testing import assert_allclose
 
 from geost import spatial
-from geost.borehole import BoreholeCollection
 from geost.utils import dataframe_to_geodataframe
